fix: changes AccountColors to AccountColor
