+++ packages/shared/lib/typings/chart.ts
@@ -1,6 +1,6 @@
 import { HistoryDataProps } from './market'
 import { AvailableExchangeRates } from './currency'
-import { AccountColors } from '../wallet'
+import { AccountColor } from './account'
 
 export interface Tooltip {
     title: string
@@ -12,7 +12,7 @@ export interface ChartData {
     label?: string
     labels?: string[]
     tooltips?: Tooltip[]
-    color?: string | AccountColors
+    color?: string | AccountColor
     steppedLine?: boolean
 }
